Don’t compile the grammar at build-time
Committed the generated parser (n1ql.cc) into the repo so that it is
no longer necessary to run the `leg` tool to build. But if the grammar
file (n1ql.leg) is modified, the cc file will need to be regenerated.

Also moved the parser code into the C++ litecore::n1ql namespace.
Not necessary for the tool, but good for when/if we move it into
LiteCore itself.

Note: n1ql.cc may trigger some compiler warnings such as unreachable
code; these should be disabled for that file in the build script.