Reworked ArgumentTokenizer
Added a `restOfInput()` method, to get the remainder of the input line
without tokenization -- this is needed so N1QL (or JSON) queries can
be entered without having to quote them.
This required reworking the entire implementation so that it doesn't
chop up the input line in advance. Instead it now reads arguments out
of the line one at a time.