Made the arguments for passing scope and collection more consistent
Side effect is that MvCommand can no longer rename documents when moving them.  The possibilities and how to explain them just got too cumbersome because it would look something like this:

Both arguments can be fully qualified document paths, unless the former contains wildcards in which case specifying a doc ID on the second argument is an error.  Both can be in two part form <part1>/<part2> in which the first argument means "<collection>/<doc>" but the second means "<scope>/<collection>" or they can be single strings in which they mean "doc" or "collection" respectively.