Updated CollectionSpec class to LiteCore's CollectionName
So, first I implemented CollectionSpec in this repo, to deal with
ownership of the slices. Then later I copied it into LiteCore, as
CollectionName, and updated it and fixed bugs.
Now I'm removing the original, both to avoid duplication and because
it still had some of those bugs. Mostly this just involved renaming.